organization + prelim. work on resource generation
Organized some scripts into folders, basic resource stuff.
Put a resource script onto prefab - can remove later.

Been working in the system (not system_test) scene.
You should be able to tap on a placeholder planet (it's a button) to
collect a resource and check out some of the input in console for when
it's available (or not).