Click+highlight star, directional buttons for touch
Click on a star to highlight. Selected star's identifier is available in
the GameManager (which is a persistent singleton). At the moment it is
possible (need to be fixed) for multi-touch to select more than one star
at a time, and for stars under the UI arrow elements to be selected.

Directional buttons for touch navigation in Sector scene. Tap
Left/Right/Up/Down to move around the zoomed out sector map in that
direction.

Altered SpeckStar prefab to have halo.