WIP: basic resource generation, organization
basic resource generation based on planet speed. power is just
constantly at the value it's supposed to be.
can collect spacebux when ready - get a debug message saying it's not
ready when it isn't time.
population auto-increments.

put planet, resource and star scripts into another folder because it was
getting messy. hope it doesn't break when merging back to master...